refactor: 토큰 검증해주는 객체 패키지 변경
