+++ src/main/java/woowacourse/auth/config/LoginInterceptor.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.config;
+package woowacourse.auth.config;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
