+++ src/test/java/woowacourse/auth/ui/AuthControllerTest.java
@@ -15,8 +15,8 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import woowacourse.auth.application.AuthService;
+import woowacourse.auth.config.WebConfig;
 import woowacourse.auth.dto.TokenRequest;
-import woowacourse.shoppingcart.config.WebConfig;
 import woowacourse.shoppingcart.exception.bodyexception.InvalidLoginException;
 
 @WebMvcTest(controllers = {AuthController.class})
