+++ src/test/java/woowacourse/shoppingcart/config/WebConfigTest.java
@@ -4,7 +4,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static woowacourse.shoppingcart.config.WebConfig.ALLOWED_METHOD_NAMES;
+import static woowacourse.auth.config.WebConfig.ALLOWED_METHOD_NAMES;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
